+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-17 22:46:31
-// Commit: feat(state): integrate custom telemetry logger in state
+// Updated at: 2026-05-18 16:05:26
+// Commit: feat(state): integrate offline caching for state state
 
 class stateComponentState {
-  final String commitRef = '1229';
-  final String lastMessage = 'feat(state): integrate custom telemetry logger in state';
-  final DateTime timestamp = DateTime.parse('2026-05-17T22:46:31+05:00');
+  final String commitRef = '1258';
+  final String lastMessage = 'feat(state): integrate offline caching for state state';
+  final DateTime timestamp = DateTime.parse('2026-05-18T16:05:26+05:00');
 }